tools/gdb: fix thread command not working
VELAPLATFO-43662

`thread` is a GDB prefix command. Use `define` can only change it to a
user prefix command. In this case, `thread 3` is unable to pass
the argument 3 to python.

Use python code to register command to overwrite this behavior. It may
not work with future GDB, but all is good for now.

Change-Id: Ic1c454c5c67522f05c751c7b7ec067ba16601e57
Signed-off-by: xuxingliang <xuxingliang@xiaomi.com>